Cleaned up WPFPlugin example a bit.
Added reference to TeklaFusion nuget package, to import theme, as is done in FusionExample repo - it makes the dialog in the designer appear as it will in TS. This required targeting to .Net Framework v4.5.1 (was v4)

Changed the MainWindow.xaml to be laid out as WPF is normally, instead of as the dsigner auto-generates. This also demonstrates a bit more of how WPFs layout components work and can be used.